Densityfitting - enforce similarity increase with adaptive force scaling
Previous to this patch, the force constant for density guided
simulations with adaptive force constant scaling was scaled up the same
amount that it was scaled down, not resulting in the announced enforced
increase in similarity. Now, the force constant is increased more strongly
than decreased, resulting in the desired enforced increase in similarity.

Change-Id: I5b149c79d91ea9e5ceeb5b0e7dce854bd6f34711