Deduplicate gmxapi documentation.
`docs/gmxapi` seems to have settled down and does not have substantial
layout changes from the documentation in python_packaging/documentation.
Compartmentalized doc builds, e.g. with docs.dockerfile, still warrant
a separate simple conf.py, so that much is retained. Otherwise, it looks
like documentation for the standalone gmxapi package or development can
be maintained in a unified way with the GROMACS project documentation,
so this change removes the transitionally duplicated content.

* Remove most of python_packaging/documentation
* Update python_packaging/docker/docs.dockerfile

Refs #2698
Refs #2985

Change-Id: I0c23f6526894ec1eec4e910463c8c9d08a7315f6