Fix .gro large box writing
Contrary to the format specification in the manual, the box components
in the .gro file were not separated by whitespace when components
(except for the first one) were >= 1000 or <= -100. Now a space is
always printed. This might not match the original GROMOS format,
but most codes reading .gro file assume space separation.
Note formatting of boxes that did have spacing is not changed.

Fixes #3176

Change-Id: I8cfcde98b46550b16b9f6bc299dc0b67454b2437