Add gmxapi_extension tests to 'tests' target.
Improve GMX_PYTHON_PACKAGE convenience targets. When GMX_PYTHON_PACKAGE
is ON, the subprojects in the python_packaging subdirectory are built.
When testing is enabled, the sample_restraint C++ unit tests were
transitively included in the 'check' target, but no mechanism was in
place to ensure they were built before the CTests and google tests
were run. This change adds the test binaries as dependencies of the
'tests' target so that the test binaries are built through normal
machinery when GROMACS is built with testing enabled and
GMX_PYTHON_PACKAGE=ON.

Change-Id: I751655dc6b6d4d58b5bbb9f5f522e26382b4f0a2