Fix mdp parameter handling in mdrun tests
Some mdp options that we intended to be set were in fact not being set.
In the two cases where mdp options were not in fact being set the default
option was desired, so this patch does not change any behavior. A note is
added to the mdp options handling code. A better long term solution might
look like popping options off the parameter map and throw if there are
options left after all defaults have been printed. This however may not
be worth the effort as the mdp options handling could change significantly
in the short term.

Change-Id: Ied6580683c91938e577861be43841e7dc3dcd8cd