Fix bug introduced by clang-format reformat.
In the post-beta2 application of clang-format, an extra closing brace
was inserted at the end of an assignment of a multi-line lambda. The
extra closing brace prematurely ended an `if` block, causing compiler
errors through the rest of the file. Simply reverting the formatting
of this section of the file does not seem to retrigger the reformatting
error.

Change-Id: I3ae83789337bb60284aeaad15eec36eb99f33708