Fix icc-19 warning in nightly matrix
The nessesity of having the default return statement after the
if-else constructions depended on the #ifdef statement. As the result,
the compiler may or may not require having the default return depending
on the definition of the GMX_LIB_MPI macros. This change moves one of
the conditional return statements outside the #ifdef construction, making
the default return statement irrelevant unconditionally of the macros
definition.

Change-Id: I1da1fe841b799f716db74fc67a8fb633af8f95d0