Allow overlapping CPU force H2D with compute
The reduction orchestration code already uses explicit sync event
in all cases and StateGpu implements the ability to schedule force
H2D in a separate stream for the "All" locality.
Hence, this change switches for non-DD runs the CPU force H2D to be done
in the update stream to allow overlap with force work in the local
stream.

Refs #3170 #3029

Change-Id: Iceb9aac395335c062109d552d3f0289688a9c75f