Fix final configuration writing for modular simulator
Modular simulator was not writing final configuration (.gro) file.
This commit fixes this, and removes an outdated sentence in the
documentation of ITrajectoryWriterClient.
I02299d34 introduces a test to avoid similar errors in the future.

Note that the complexity required to introduce this feature implies that
trajectory writing in the modular simulator needs some work for
GROMACS 2021.

Change-Id: Ie7f03f5beea2d78411c7a0cdfa85f71deeb622fa