Finalize gmxapi 0.1
Set 0.1 release metadata for Python package and C++ library.

Note that 0.1 is API identical to 0.0.8, but we are
taking the next step to the targeted semantic versioning,
and patch releases should no longer affect the APIs.
However, we are not yet at a point where we can have
clear stability guarantees between minor versions.

Change-Id: I4f4a6ba84353730e2962b730cd1e656bae4e377a