Change grompp decoupling SD warning to note
The warning that grompp issues when decoupling a molecule
without the use of the SD integrator has been changed to a note,
since there are valid use cases for using normal MD when not
running in the completely decoupled state.

Note that checking for actual decoupled VdW states requires a lot of
code and would still not be 100% correct.

Fixes #2767

Change-Id: I55c84afcbcb7b19bdb8a7a8bc586a65c50f8d0c1