Expand ListOfLists pushBack() functionality
Added member pushBackListOfSize(int) to ListOfLists to add a list of a
given size to the list so its elements can be set. Used this in
applicable cases and added back() to ListOfLists to facilitate access
to that newly added list. Added front() for symmetry. Added tests.

Limited ListOfLists to arithmetic types, as there is no compelling
use case for other types and it simplifies the class.

Also fixed two minor issues that came up during review.

Change-Id: Ia05c270f9c337902c78f97e9613e45e82a21df63