Add a note to the log file when -update gpu was set with DD
GPU update with DD is not enabled with DD even if '-update gpu'
is requested. The only way to enable it is to set
GMX_FORCE_UPDATE_DEFAULT_GPU environment variable. This patch
adds a note to the output, expaining that. Also, extra warning
note is added when the environment variable is set.

Fixes #3259.

Change-Id: I480dd8e734ff60156e117cc80b1cd9af6a161791