Use `todo` Sphinx directive more.
Start making "to do"s more actionable. Use the Sphinx `todo` directive
to allow special processing. For comment blocks that imply planned
activity, convert comment to `todo` directive for more easily discovered
and tracked doctree elements.

"To do" items should have a clear call to action and some follow through,
such as being prohibited in release branches or at least referring to
tracked issues, but that is beyond the scope of this change.
This change only seeks to make the mass of "to do"s more discoverable.

Change-Id: I8cbc06a5a1d78791bc6e9b0a2530632ad09a9380