Remove GMX_CYCLECOUNTERS
With ARMv7 support removed, cycle counters are available on all
known platforms, so the CMake option to disable them can be removed.
This change will be easy to revert if the need arises in future.

Includes partial revert of 0d4ea6033cba8cbfe6, which was added
when supporting quirks of ARMv7.

Refs #2990

Change-Id: Iac107d977fcd8ea67d8306149b9fb6c17985afa4