Tests that verify serializer behaviours
xdr_vector serializing char uses four bytes per char, which is
wasteful when that char buffer is just a binary dump such as with the
tpr file body. This test ensures that our assumption when writing tpr
files is that we can reinterpret as a wider data type and avoid excess
padding.

It also tests that xdr_opaque does not introduce such padding.

Refs #3269
Refs #3276

Change-Id: I9c40b8ab7b9dabe461cb4dd146bb132d8f5a9681