Remove stray "to do".
The context of the comment has changed over several commits. If it is
still relevant, its meaning is unclear. If there is a question that still
needs to be resolved, it can be re-posed in the issue tracking system
when someone has a chance to clarify and frame the discussion.

Change-Id: Ifb65ae4942a1225aa23189c95aaa4c73c3b7c88c