Make nbnxm headers more self-contained
Resolves existing TODO, removing a dependency of the CPU kernels
on config.h

Eliminates an unnecessary function, and the inlining of a function
that is only called every NS step.

Renamed gmx_nbnxn_*_t to gmx_nbnxm_gpu_t because there was no
advantage to the previous scheme. Each GPU build configuration only
ever uses one type, so there is no problem having different
declarations in different build configurations. The compiler would
complain if there was. The renaming of the former collection of types
in gpu_types.h created a dependency on config.h which should be
avoided.

Something about the renaming of gmx_nbnxm_cuda_t broke the way Doxygen
understood nbnxm_cuda_types.h. I thought that declaring a Doxygen
header for atomdata.h had fixed it, somehow, but it didn't, so there
is a suppression for the suspected false positive. Also fixed the
Doxygen in atomdata.h

Change-Id: I18c573ad87cbab9ae4f38aa49541f0821dc1c145