Merge VV implementations of ComputeGlobalsElement
Until now, ComputeGlobalsElement had effectively two separate
(templated) implementations for velocity verlet, to be called
at full timestep and after the position update.
As we gradually move towards a design in which the
ComputeGlobalsElement is agnostic to the data it reduces, this
separation does not make sense anymore. As a first step, the
two implementations are therefore merged.

Change-Id: If0b549a3d87fea50e895b70c55632e1f58582067