Make NbnxnGpu class with constructor
This is needed for future reform of the NB GPU code. Also converted
the "staging" struct to have initializers (used to make the default
constructor) because it is held by value and would otherwise have
default constructor that leaves fields uninitialized.

Also replaced cl_bool with normal bool, as there was no advantage for
the former.

Change-Id: I1b63f0a8145dbd911062d1039f129074b0abdc3e