Fixed bug when fitting zeta.
The screened width of Gaussians and Slaters can be optimized by
tune_eem. For this the ACM charge generation needs the updated
zeta, but in case of shell models the GROMACS code also needs to
be updated. The latter part was missing and therefore the relax_shells
routine only ever used the original zeta, leading to irreproducible
results.

Change-Id: I9ce489095c7527532fc041b09565edc29e4e5568