Actually fix PME forces with FE without perturbed q/LJ
PME would incorrectly ignore the mesh forces on perturbed atoms when
no charges or LJ atom types were actually perturbed.
This is because the "fix" in commit 12f01f79 incorrectly changed the
lambda value used for the PME forces from the actual lambda value
to 1 instead of to 0 as it should have been.

Fixes #2640
Fixes #3359

Change-Id: Ib19565bf017ddbd2fddc66c52be0ff9efc7e707c