Consolidate test systems into library
This change moves the client code used for testing to a central
location instead of duplicating code in several test files. It
also adds more complete testing for SimulationState, but comments
out nbkernelsystem tests as they are in flux. A few getters were
needed to make testing easier, and they have been added, with
related tests for those getters.