Add spc-methanol system to test systems
While not currently used, the spc-methanol system should be useful
for comparing results of nbnxm initialization with default gromacs
because this system is used for testing mdrun. Once nonbonded
kernels are working this system can also be used to test energies,
though bonds and settles would need to be added to nblib to
completely replicate gromacs.