Introduce readXvgData to read xvg data into std::vector
This allows for modernization of table functions, demonstrated
here with read_tables, as well as potentially any other function
that needs to read in xvg data. The legacy read_xvg is now
implemented in terms of readXvgData. There is already a test for
table functions and this change now adds tests for reading in
xvg data.

Change-Id: Iac0b13d7db15f04a8b0b464df9fa136ff4b2a213