Added check for -nexcl command line in gentop and tune_eem
The number of exclusions should be changed from the gentop.dat
value *only* if the used explicitly sets it on the command line,
not by default. Added checks for this in gentop.cpp and molgen.cpp.

Added other debug output to tune_eem.

Cleaned compiler warning due to obsolete code in penalty checking
in tune_eem.

Change-Id: Ib271d8ddd93ca32a41816155e93d2a14fac82a49