Fix incorrect GPU update task assignment fallback
This change fixes the incorrect fallback condition in the GPU update
task assignment. This also eliminates the assertion trigerred in do_md()
where the correct form of the check is asserted on.
Since the fallback path only triggers when the GPU update would be used by
default this only affected runs with GMX_FORCE_UPDATE_DEFAULT_GPU.

Change-Id: I8b2b506fa998b6ed5e6d6d8ae4b78e27006df199