Combine initialization routines for GPUs assigned to PME and Nonbonded
The subroutines that do the initialization share a lot of logic. Also,
the current setup does not allow to have more than one GPU assigned to
a single rank. This make it natural to keep one handle to the GPU
device, if there is a GPU assigned to the rank.

Change-Id: I9177e1a6c00cf6efa7b58370b8fa6220c8fd4496