Added more docs and tests to EnumerationArray
It is now more clear that there is no risk of a segfault if somehow
Foo::Count was used to look up a value from an EnumerationArray. The
assertion would not be nice for a user to get either, but the positive
aspects that flow from having enumeration helpers that need the Count
field outweighs this downside.

Change-Id: I6111e79042326044a807b9b8ac58f32760a56b82