Add argon forces test
After testing to make sure that the non-bonded force parameters
(nbfp) are the same in nblib and mdrun, the forces were obtained
from mdrun and found to be the same as nblib. A test with the
values from mdrun is now added. This shows that the issue with
forces for the spc and methanol system is likely due to ordering
of nbfp.