Move testing code to test directory
This should not add bulk to libgromacs

Also decoupled some headers from unnecessarily depending on other
headers, some of which inhibited the proper separation of code and
tests. In particular, there's now a proper getter for warp size.

This needed the default implementation of pmeInitEmpty to move into a
source file, because it was requiring the definition of CodePath to be
available.

Change-Id: I3040c77404ade4d8cf0a41faf4aa4da4db42e3d3