Convert t_graph to C++
Convert memory management and initialization of t_graph to C++.
Added a simple class to build the graph. This makes graph generation
a factor 4 slower, but we plan to use gmx_mtop_t for generating
the graph, which will lead to a large speed-up.

Todo: Convert t_graph to a proper class.

Change-Id: Ib5a3059e9c0d0cf6888448fe471280ecf6b9467a