Remove redundant jobs.
Remove distinction between jobs that do and do not stage a build of the
gmxapi Python package in the build tree for API breakage testing and
full documentation builds.

Set a default CMake flag of "-DGMX_PYTHON_PACKAGE=ON" and install some
Python package dependencies with the default EXTRA_INSTALLS. The extra
installs will go away in the near future with updated Docker images.

Removes (up to) 12 jobs from the pipeline.

Change-Id: I13a60b7323ac0bf240f00b3b1360bedd2886e902